Move get_conf to separate module
It much more convenient to move `get_conf` function into separate
module, in order to make it importable from tests, for example.
