+++ xsnippet_api/__main__.py
@@ -11,23 +11,11 @@
 
 import os
 import sys
-import configparser
 
 from aiohttp import web
 
 from xsnippet_api.application import create_app
-
-
-def get_conf(paths, envvar=None):
-    conf = configparser.ConfigParser()
-    conf.read(paths)
-
-    # the idea behind that thing is to provide an additional
-    # settings file with production overrides
-    if envvar in os.environ:
-        conf.read(os.environ[envvar])
-
-    return conf
+from xsnippet_api.conf import get_conf
 
 
 def main(args=sys.argv[1:]):
